--- CachedRoboCopy/CachedRoboCopy.csproj
@@ -1,32 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-      <TargetFrameworks>
-        netstandard2.0;
-        netcoreapp3.1;
-        net472;
-        net5.0-windows
-      </TargetFrameworks>
-      <OutputType>Library</OutputType>
-      <LangVersion>9</LangVersion>
-      <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
-      <RootNamespace>RFBCodeWorks.CachedRoboCopy</RootNamespace>
-  </PropertyGroup>
-    <PropertyGroup Condition="'$(Configuration)'=='Debug'">
-      <DefineConstants>DEBUG;TRACE</DefineConstants>
-      <DocumentationFile>bin\Debug\$(TargetFramework)\CachedRoboCopy.xml</DocumentationFile>
-    </PropertyGroup>
-    <PropertyGroup Condition="'$(Configuration)'=='Release'">
-      <DebugType>pdbonly</DebugType>
-      <Optimize>true</Optimize>
-      <OutputPath>bin\Release\</OutputPath>
-      <DefineConstants>TRACE</DefineConstants>
-      <ErrorReport>prompt</ErrorReport>
-      <WarningLevel>4</WarningLevel>
-      <DocumentationFile>bin\Release\$(TargetFramework)\CachedRoboCopy.xml</DocumentationFile>
-    </PropertyGroup>
-    <ItemGroup>
-      <ProjectReference Include="..\..\..\RoboSharp\RoboSharp\RoboSharp.csproj">
-        <Private>false</Private>
-      </ProjectReference>
-    </ItemGroup>
-</Project>
